+++ doc/conf.py
@@ -27,6 +27,7 @@ nb_execution_excludepatterns = [
     "integrations/oracle.ipynb",
     "integrations/snowflake.ipynb",
     "integrations/redshift.ipynb",
+    "integrations/spark.ipynb",
 ]
 nb_execution_in_temp = True
 nb_execution_show_tb = True