+++ src/sql/plot.py
@@ -268,10 +268,9 @@ FROM {{table}}
 """
     if use_backticks:
         template_ = template_.replace('"', "`")
-
+        table = table.replace('"', "`")
     template = Template(template_)
     query = template.render(table=table, column=column)
-
     min_, max_ = conn.execute(query, with_).fetchone()
     return min_, max_
 
@@ -628,6 +627,7 @@ def _histogram(
 
             if use_backticks:
                 template_ = template_.replace('"', "`")
+                table = table.replace('"', "`")
 
             template = Template(template_)
 
@@ -663,6 +663,7 @@ def _histogram(
 
             if use_backticks:
                 template_ = template_.replace('"', "`")
+                table = table.replace('"', "`")
 
             template = Template(template_)
 
@@ -681,6 +682,7 @@ def _histogram(
 
         if use_backticks:
             template_ = template_.replace('"', "`")
+            table = table.replace('"', "`")
 
         template = Template(template_)
 
@@ -835,6 +837,7 @@ def _bar(table, column, with_=None, conn=None):
 
         if use_backticks:
             template_ = template_.replace('"', "`")
+            table = table.replace('"', "`")
 
         template = Template(template_)
         query = template.render(table=table, x_=x_, height_=height_)
@@ -854,6 +857,7 @@ def _bar(table, column, with_=None, conn=None):
 
         if use_backticks:
             template_ = template_.replace('"', "`")
+            table = table.replace('"', "`")
 
         template = Template(template_)
         query = template.render(table=table, column=column)
@@ -1022,6 +1026,7 @@ def _pie(table, column, with_=None, conn=None):
                 """
         if use_backticks:
             template_ = template_.replace('"', "`")
+            table = table.replace('"', "`")
 
         template = Template(template_)
         query = template.render(table=table, labels_=labels_, size_=size_)
@@ -1037,6 +1042,7 @@ def _pie(table, column, with_=None, conn=None):
                 """
         if use_backticks:
             template_ = template_.replace('"', "`")
+            table = table.replace('"', "`")
 
         template = Template(template_)
         query = template.render(table=table, column=column)