+++ src/tests/test_connection.py
@@ -13,6 +13,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.engine import Engine
 from sqlalchemy import exc
 
+
 from sql.connection import connection as connection_module
 import sql.connection
 from sql.connection import (
@@ -21,6 +22,7 @@ from sql.connection import (
     ConnectionManager,
     is_pep249_compliant,
     default_alias_for_engine,
+    is_spark,
     ResultSetCollection,
     detect_duckdb_summarize_or_select,
 )
@@ -39,6 +41,34 @@ def mock_database(monkeypatch, cleanup):
     monkeypatch.setitem(sys.modules, "some_driver", Mock())
     monkeypatch.setattr(Engine, "connect", Mock())
     monkeypatch.setattr(sqlalchemy, "create_engine", Mock())
+
+
+def mock_sparksession():
+    mock = Mock(
+        spec=[
+            "table",
+            "read",
+            "createDataFrame",
+            "sql",
+            "stop",
+            "catalog",
+            "version",
+        ]
+    )
+    return mock
+
+
+def mock_not_sparksession():
+    mock = Mock(
+        spec=[
+            "read",
+            "readStream",
+            "createDataFrame",
+            "sql",
+            "version",
+        ]
+    )
+    return mock
 
 
 @pytest.fixture
@@ -457,6 +487,24 @@ def test_is_pep249_compliant(conn, expected):
     assert is_pep249_compliant(conn) is expected
 
 
+@pytest.mark.parametrize(
+    "descriptor, expected",
+    [
+        [sqlite3.connect(""), False],
+        [duckdb.connect(""), False],
+        [create_engine("sqlite://"), False],
+        [mock_sparksession(), True],
+        [mock_not_sparksession(), False],
+        [None, False],
+        [object(), False],
+        ["not_a_valid_connection", False],
+        [0, False],
+    ],
+)
+def test_is_spark(descriptor, expected):
+    assert is_spark(descriptor) is expected
+
+
 def test_close_all(ip_empty, monkeypatch):
     connections = {}
     monkeypatch.setattr(ConnectionManager, "connections", connections)
@@ -585,6 +633,22 @@ def test_set_dbapi(monkeypatch, callable_, key):
     monkeypatch.setattr(ConnectionManager, "connections", connections)
 
     conn = ConnectionManager.set(callable_(""), displaycon=False)
+
+    assert connections == {key: conn}
+    assert ConnectionManager.current == conn
+
+
+@pytest.mark.parametrize(
+    "spark, key",
+    [
+        [mock_sparksession(), "Mock"],
+    ],
+)
+def test_set_spark(monkeypatch, spark, key):
+    connections = {}
+    monkeypatch.setattr(ConnectionManager, "connections", connections)
+
+    conn = ConnectionManager.set(spark, displaycon=False)
 
     assert connections == {key: conn}
     assert ConnectionManager.current == conn